tools/requirements.txt
instaloader
Pillow
pytesseract
openai-whisper
torch
requests
python-dotenv